feat(rust): optimize serializers, ObjectGraph, and module split
Reuse encode buffers and prost encoded_len, support ObjectGraph via a
flat node table for all codecs, and split wrappers into family modules
aligned with Go/Python/C layout. Includes unit tests and docs updates.
